Time: 50 ms (29.89%), Space: 15.9 MB (92.46%) - LeetHub

+++ 437-path-sum-iii/437-path-sum-iii.cpp
@@ -14,7 +14,7 @@ typedef long long int lli;
 class Solution {
 public:
     
-    lli pathSum(TreeNode* root, long long int targetSum) {
+    lli pathSum(TreeNode* root, lli targetSum) {
         if (root == NULL) return 0;
         
         lli count = 0;
